[consensus] block metadata uses a bitmap of validators
Convert block metadata to use a bitmap of validators to denote voters
instead of the previous list of voters. This will make the block metadata
smaller; it will be especially noticeable with a large set of validators
and a small number of transactions.

This change deprecates the existing previous_block_votes column in indexer.
The new column previous_block_votes_bitmap contains the information as a
bitmap. For users of indexer (e.g., explorer) to map this to the actual
addresses, the sorted validator list per epoch is also needed. This is
future work.

Closes: #712